refactor(client): :recycle: move components
