+++ client/src/containers/Backend/UserTranslation/UserTranslation.module.scss
@@ -1,3 +1,14 @@
+@import "src/scss/colors";
+
+.main {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  background-color: $gray20;
+  margin-top: -95px;
+  padding-top: 100px;
+}
+
 .container {
   display: flex;
   flex: 1;
