+++ client/src/containers/Backend/UserTranslation/components/OneNeedTranslationModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./OneNeedTranslationModal";
