+++ client/src/containers/Backend/UserTranslation/components/TranslationLanguagesChoiceModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TranslationLanguagesChoiceModal";
