+++ client/src/containers/Backend/UserTranslation/components/TranslationNeedsList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TranslationNeedsList";
