+++ client/src/containers/Backend/UserTranslation/components/TranslationsAvancement/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TranslationsAvancement";
