+++ client/src/containers/Backend/UserTranslation/components/__tests__/functions.test.ts
@@ -1,5 +1,5 @@
 // @ts-nocheck
-import { filterData, sortData } from "../functions";
+import { sortData } from "../functions";
 const data1 = {
   tradStatus: "À traduire",
   typeContenu: "dispositif",
@@ -68,89 +68,6 @@ const data7 = {
 };
 
 const data = [data1, data2, data3, data4, data5, data6, data7, data8];
-describe("filter data", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "all", true, "all", "");
-    expect(result).toEqual({
-      dataToDisplay: data,
-      nbARevoir: 2,
-      nbATraduire: 2,
-      nbAValider: 2,
-      nbPubliees: 2,
-      nbDispositifs: 4,
-      nbDemarches: 4,
-    });
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "all", false, "all", "");
-    expect(result).toEqual({
-      dataToDisplay: [data1, data2, data6, data7],
-      nbARevoir: 0,
-      nbATraduire: 2,
-      nbAValider: 0,
-      nbPubliees: 2,
-      nbDispositifs: 2,
-      nbDemarches: 2,
-    });
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "À revoir", true, "all", "");
-    expect(result).toEqual({
-      dataToDisplay: [data3, data4],
-      nbARevoir: 2,
-      nbATraduire: 2,
-      nbAValider: 2,
-      nbPubliees: 2,
-      nbDispositifs: 1,
-      nbDemarches: 1,
-    });
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "Validée", false, "dispositif", "");
-    expect(result).toEqual({
-      dataToDisplay: [data6],
-      nbARevoir: 0,
-      nbATraduire: 2,
-      nbAValider: 0,
-      nbPubliees: 2,
-      nbDispositifs: 1,
-      nbDemarches: 1,
-    });
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "all", false, "dispositif", "tif6");
-    expect(result).toEqual({
-      dataToDisplay: [data6],
-      nbARevoir: 0,
-      nbATraduire: 0,
-      nbAValider: 0,
-      nbPubliees: 1,
-      nbDispositifs: 1,
-      nbDemarches: 0,
-    });
-  });
-
-  it("should return correct data ", () => {
-    const result = filterData(data, "all", true, "dispositif", "Marqué");
-    expect(result).toEqual({
-      dataToDisplay: [data1, data3, data6, data8],
-      nbARevoir: 1,
-      nbATraduire: 1,
-      nbAValider: 1,
-      nbPubliees: 1,
-      nbDispositifs: 4,
-      nbDemarches: 0,
-    });
-  });
-});
 
 describe("sortData", () => {
   it("should sort data", () => {
