+++ client/src/containers/Backend/UserTranslation/components/index.ts
@@ -1,0 +1,11 @@
+import OneNeedTranslationModal from "./OneNeedTranslationModal";
+import TranslationLanguagesChoiceModal from "./TranslationLanguagesChoiceModal";
+import TranslationNeedsList from "./TranslationNeedsList";
+import TranslationsAvancement from "./TranslationsAvancement";
+
+export {
+  OneNeedTranslationModal,
+  TranslationLanguagesChoiceModal,
+  TranslationNeedsList,
+  TranslationsAvancement
+}
